# BE/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LingoPrep API",
    "description": "Intelligent IELTS & Aptis Speaking & Writing Practice & Assessment Platform powered by AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/questions": {
      "get": {
        "tags": [
          "Question Bank"
        ],
        "summary": "Get Questions",
        "operationId": "get_questions_api_questions_get",
        "parameters": [
          {
            "name": "exam_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExamType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exam Type"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SkillType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skill"
            }
          },
          {
            "name": "part",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Part"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Get Questions Api Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Question Bank"
        ],
        "summary": "Create Question",
        "operationId": "create_question_api_questions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}": {
      "get": {
        "tags": [
          "Question Bank"
        ],
        "summary": "Get Question",
        "operationId": "get_question_api_questions__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Question Bank"
        ],
        "summary": "Update Question",
        "operationId": "update_question_api_questions__question_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Question Bank"
        ],
        "summary": "Delete Question",
        "operationId": "delete_question_api_questions__question_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews/queue": {
      "get": {
        "tags": [
          "Teacher Review Module"
        ],
        "summary": "Get Review Queue",
        "description": "Returns submissions that need teacher review or have been completed recently.\nPrioritizes REVIEW_REQUESTED submissions.",
        "operationId": "get_review_queue_api_reviews_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponse"
                  },
                  "type": "array",
                  "title": "Response Get Review Queue Api Reviews Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/reviews/{submission_id}": {
      "post": {
        "tags": [
          "Teacher Review Module"
        ],
        "summary": "Submit Teacher Review",
        "operationId": "submit_teacher_review_api_reviews__submission_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamType": {
        "type": "string",
        "enum": [
          "IELTS",
          "APTIS"
        ],
        "title": "ExamType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionCreate": {
        "properties": {
          "exam_type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "part": {
            "type": "string",
            "title": "Part"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "default": "General"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty",
            "default": "Medium"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "prep_time_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Time Seconds",
            "default": 60
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds",
            "default": 120
          },
          "min_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Words"
          },
          "max_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Words"
          },
          "sample_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Answer"
          },
          "rubric_criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Criteria"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "default": []
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "exam_type",
          "skill",
          "part",
          "title",
          "prompt"
        ],
        "title": "QuestionCreate"
      },
      "QuestionResponse": {
        "properties": {
          "exam_type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "part": {
            "type": "string",
            "title": "Part"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "default": "General"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty",
            "default": "Medium"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "prep_time_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Time Seconds",
            "default": 60
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds",
            "default": 120
          },
          "min_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Words"
          },
          "max_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Words"
          },
          "sample_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Answer"
          },
          "rubric_criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Criteria"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "default": []
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "exam_type",
          "skill",
          "part",
          "title",
          "prompt",
          "id",
          "created_at"
        ],
        "title": "QuestionResponse"
      },
      "QuestionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          },
          "min_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Words"
          },
          "max_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Words"
          },
          "sample_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Answer"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "QuestionUpdate"
      },
...
      "SkillType": {
        "type": "string",
        "enum": [
          "SPEAKING",
          "WRITING"
        ],
        "title": "SkillType"
      },
...
      "SubmissionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "submission_type": {
            "$ref": "#/components/schemas/SkillType"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "audio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Path"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "teacher_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeacherReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "question_id",
          "submission_type",
          "duration_seconds",
          "word_count",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SubmissionResponse"
      },
...
      "TeacherReviewCreate": {
        "properties": {
          "overall_band": {
            "type": "number",
            "title": "Overall Band"
          },
          "overall_cefr": {
            "type": "string",
            "title": "Overall Cefr"
          },
          "criteria_scores": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria Scores"
          },
          "teacher_notes": {
            "type": "string",
            "title": "Teacher Notes"
          }
        },
        "type": "object",
        "required": [
          "overall_band",
          "overall_cefr",
          "teacher_notes"
        ],
        "title": "TeacherReviewCreate"
      },
      "TeacherReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "teacher_id": {
            "type": "integer",
            "title": "Teacher Id"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "overall_band": {
            "type": "number",
            "title": "Overall Band"
          },
          "overall_cefr": {
            "type": "string",
            "title": "Overall Cefr"
          },
          "criteria_scores": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria Scores"
          },
          "teacher_notes": {
            "type": "string",
            "title": "Teacher Notes"
          },
          "is_overridden": {
            "type": "boolean",
            "title": "Is Overridden"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "submission_id",
          "teacher_id",
          "overall_band",
          "overall_cefr",
          "teacher_notes",
          "is_overridden",
          "reviewed_at"
        ],
        "title": "TeacherReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  },
...
}